pyproject.toml
[tool.poetry]
name = "logtools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pytz = "^2023.3.post1"
colored = "^2.2.3"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
ipython = "^8.15.0"
mypy = "^1.6.0"
install = "^1.3.5"
types-pytz = "^2023.3.1.1"
types-python-dateutil = "^2.8.19.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
log-merge = 'logtools.cli.merge:main'